requirements.txt
flask>=3.0.0
gunicorn>=21.0.0
retina-face>=0.0.17
deepface>=0.0.93
tf-keras>=2.15.0
pillow>=10.0.0
numpy>=1.24.0
opencv-python-headless>=4.8.0
